+++ src/components/icons/ContactsIcon.tsx
@@ -1,11 +1,7 @@
-import GithubIcon from './icons/GithubIcon';
-import MailIcon from './icons/MailIcon';
-import TagIcon from './icons/TagIcon';
-import VelogIcon from './icons/VelogIcon';
-
-type ContextType = {
-  contact: 'email' | 'github' | 'velog';
-};
+import GithubIcon from './GithubIcon';
+import MailIcon from './MailIcon';
+import TagIcon from './TagIcon';
+import VelogIcon from './VelogIcon';
 
 const icons: { [key: string]: React.FunctionComponent } = {
   email: MailIcon,
