+++ src/pages/blog/index.tsx
@@ -1,11 +1,8 @@
 import { AnimatePresence, motion } from 'framer-motion';
 import Link from 'next/link';
 
-import PlainText from '@/common/PlainText';
-import SubTitle from '@/common/SubTitle';
-import Title from '@/common/Title';
-import HoverCard from '@/components/HoverCard';
-import PostList from '@/components/PostList';
+import { HoverCard, PlainText, SubTitle, Title } from '@/components/common';
+import PostList from '@/components/post/PostList';
 import { PageSEO } from '@/components/SEO';
 import {
   fadeIn,
